Restructured the code, added actionCreator file, add new functionalities to the todo list
