+++ todo/src/index.js
@@ -2,8 +2,9 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { createStore, compose } from 'redux';
 import { Provider } from 'react-redux';
+import { BrowserRouter as Router } from 'react-router-dom';
 
-import reducer from './store/reducer';
+import reducer from './store/reducers/reducer';
 
 import './index.css';
 import App from './App';
@@ -12,8 +13,10 @@ const store = createStore(reducer,  compose( window.__REDUX_DEVTOOLS_EXTENSION__
 
 
 ReactDOM.render(
-    <Provider store={store}>
-        <App />
-    </Provider>,
+    <Router>
+        <Provider store={store}>
+            <App />
+        </Provider>
+    </Router>,
     document.getElementById('root')
 );
